working version of the code that sets up the bluetooth server and allows devices to connect and subscribe. audio transmission has not been tested yet

+++ transmitter_tag/main/adpcm.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 typedef struct {
-    int16_t predicted;
+    int32_t predicted;
     int     step_index;
 } adpcm_state_t;
